refactored code for calendar view
